pyproject.toml
[project]
name = "garagedoor-opener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
]
